+++ assets/radar-rules.js
@@ -830,22 +830,6 @@
             { title: '小组-精华', docs: 'https://docs.rsshub.app/social-media.html#dou-ban', source: '/group/:groupid', target: '/douban/group/:groupid/elite' },
         ],
     },
-    'okjike.com': {
-        _name: '即刻',
-        m: [
-            {
-                title: '用户动态',
-                docs: 'https://docs.rsshub.app/social-media.html#ji-ke-yong-hu-dong-tai',
-                source: '/reposts/:repostId',
-                target: (params, url, document) => {
-                    const uid = document.querySelector('.avatar').getAttribute('href').replace('/users/', '');
-                    return uid ? `/jike/user/${uid}` : '';
-                },
-            },
-            { title: '圈子', docs: 'https://docs.rsshub.app/social-media.html#ji-ke-quan-zi', source: '/topics/:id', target: '/jike/topic/:id' },
-            { title: '圈子 - 纯文字', docs: 'https://docs.rsshub.app/social-media.html#ji-ke-quan-zi-chun-wen-zi', source: '/topics/:id', target: '/jike/topic/text/:id' },
-        ],
-    },
     'ems.com.cn': { _name: '中国邮政速递物流', www: [{ title: '新闻', docs: 'https://docs.rsshub.app/other.html#zhong-guo-you-zheng-su-di-wu-liu', source: '/aboutus/xin_wen_yu_shi_jian.html', target: '/ems/news' }] },
     'popiapp.cn': {
         _name: 'Popi 提问箱',
