+++ lib/router.js
@@ -2904,10 +2904,10 @@ router.get('/futunn/highlights', lazyloadRouteHandler('./routes/futunn/highlight
 // 外接大脑
 router.get('/waijiedanao/article/:caty', lazyloadRouteHandler('./routes/waijiedanao/article'));
 
-// 即刻
-router.get('/jike/topic/:id', lazyloadRouteHandler('./routes/jike/topic'));
-router.get('/jike/topic/text/:id', lazyloadRouteHandler('./routes/jike/topicText'));
-router.get('/jike/user/:id', lazyloadRouteHandler('./routes/jike/user'));
+// 即刻 migrated to v2
+// router.get('/jike/topic/:id', lazyloadRouteHandler('./routes/jike/topic'));
+// router.get('/jike/topic/text/:id', lazyloadRouteHandler('./routes/jike/topicText'));
+// router.get('/jike/user/:id', lazyloadRouteHandler('./routes/jike/user'));
 
 // 网易新闻
 // router.get('/netease/news/rank/:category?/:type?/:time?', lazyloadRouteHandler('./routes/netease/news/rank'));
