+++ lib/v2/jike/common.js
@@ -1,4 +1,3 @@
-const url = require('url');
 const { parseDate } = require('@/utils/parse-date');
 
 module.exports = {
@@ -43,7 +42,7 @@ module.exports = {
 
                 // 对于即刻抓取的微信公众号文章 特殊处理
                 // 此时 Rss原文链接 变为 微信公众号链接
-                if (url.parse(linkUrl).host === 'mp.weixin.qq.com') {
+                if (new URL(linkUrl).host === 'mp.weixin.qq.com') {
                     link = linkUrl;
                 }
