+++ lib/v2/jike/maintainer.js
@@ -1,0 +1,5 @@
+module.exports = {
+    '/topic/text/:id': ['HenryQW'],
+    '/topic/:id': ['DIYgod', 'prnake'],
+    '/user/:id': ['DIYgod', 'prnake'],
+};
