+++ lib/v2/jike/router.js
@@ -1,0 +1,5 @@
+module.exports = (router) => {
+    router.get('/topic/text/:id', require('./topicText'));
+    router.get('/topic/:id', require('./topic'));
+    router.get('/user/:id', require('./user'));
+};
