+++ lib/v2/jike/topic.js
@@ -26,16 +26,13 @@ module.exports = async (ctx) => {
         item: common.topicDataHanding(data),
     };
     if (id === '553870e8e4b0cafb0a1bef68' || id === '55963702e4b0d84d2c30ce6f') {
-        const promises = ctx.state.data.item.map(async (one) => {
+        const promises = ctx.state.data.item.map((one) => {
             const item = { ...one };
             const regResult = /https:\/\/www.okjike.com\/medium\/[a-zA-Z0-9]*/.exec(item.description);
             if (regResult) {
                 const newsUrl = regResult[0];
-                const cache = await ctx.cache.get(newsUrl);
-                if (cache) {
-                    item.description = cache;
-                } else {
-                    const { data } = await got.get(newsUrl);
+                item.description = ctx.cache.tryGet(newsUrl, async () => {
+                    const { data } = await got(newsUrl);
                     const $ = cheerio.load(data);
                     const upper = $('ul.main > li.item');
                     const links = upper.find('a').map((_, ele) => $(ele).attr('href'));
@@ -44,11 +41,8 @@ module.exports = async (ctx) => {
                     for (let i = 0; i < links.length; i++) {
                         description += `${i + 1}、<a href="${links[i]}">${texts[i]}</a><br>`;
                     }
-                    if (description) {
-                        item.description = description;
-                        await ctx.cache.set(newsUrl, description);
-                    }
-                }
+                    return description;
+                });
             }
             item.title = `${topic.content} ${dayjs(one.pubDate).format('MM月DD日')}`;
             return item;
