+++ lib/v2/jike/user.js
@@ -134,7 +134,7 @@ module.exports = async (ctx) => {
                     });
             }
 
-            return Promise.resolve(single);
+            return single;
         })
     );
